requirements.txt
firebase_admin
requests
fastapi
uvicorn
python-dotenv
